test: cover batched worker port discovery
