+++ tests/unit/distributed/test_virtual_cluster.py
@@ -20,7 +20,6 @@ from unittest.mock import MagicMock, patch
 import pytest
 import ray
 
-import nemo_rl.distributed.virtual_cluster as virtual_cluster
 from nemo_rl.distributed.virtual_cluster import (
     DEFAULT_MASTER_PORT_RANGE_HIGH,
     DEFAULT_MASTER_PORT_RANGE_LOW,
@@ -315,101 +314,6 @@ class TestGetFreePortLocal:
         }
         # With a range of 800 ports, 10 calls should very likely produce multiple unique ports
         assert len(ports) > 1
-
-
-class TestRayVirtualClusterBatchPorts:
-    """Tests for batched address and port discovery."""
-
-    class FakePortFinder:
-        def __init__(self):
-            self.refs = []
-
-        def options(self, **kwargs):
-            return self
-
-        def remote(self, port_range_low, port_range_high):
-            ref = f"ref-{len(self.refs)}"
-            self.refs.append((ref, port_range_low, port_range_high))
-            return ref
-
-    @staticmethod
-    def _make_cluster():
-        cluster = RayVirtualCluster(bundle_ct_per_node_list=[2, 1], use_gpus=False)
-        cluster._node_placement_groups = [
-            MagicMock(bundle_specs=[{"CPU": 1}]),
-            MagicMock(bundle_specs=[{"CPU": 1}]),
-        ]
-        return cluster
-
-    @staticmethod
-    def _patch_common_mocks(monkeypatch, port_finder, results_by_ref):
-        def fake_get(refs):
-            return [results_by_ref[ref] for ref in refs]
-
-        monkeypatch.setattr(virtual_cluster, "_get_node_ip_and_free_port", port_finder)
-        monkeypatch.setattr(
-            virtual_cluster,
-            "PlacementGroupSchedulingStrategy",
-            lambda **kwargs: kwargs,
-        )
-        monkeypatch.setattr(virtual_cluster.ray, "get", fake_get)
-
-    def test_default_batch_size_uses_ray_get_fast_path(self, monkeypatch):
-        cluster = self._make_cluster()
-        port_finder = self.FakePortFinder()
-        results_by_ref = {
-            "ref-0": ("node-a", 25001),
-            "ref-1": ("node-a", 25002),
-            "ref-2": ("node-b", 25003),
-        }
-        self._patch_common_mocks(monkeypatch, port_finder, results_by_ref)
-        wait_mock = MagicMock(
-            side_effect=AssertionError("fast path should not call ray.wait")
-        )
-        monkeypatch.setattr(virtual_cluster.ray, "wait", wait_mock)
-
-        results = cluster.get_available_addresses_and_ports_batch(
-            [(0, 0), (0, 1), (1, 0)]
-        )
-
-        assert results == [
-            ("node-a", 25001),
-            ("node-a", 25002),
-            ("node-b", 25003),
-        ]
-        assert wait_mock.call_count == 0
-
-    def test_preserves_input_order_when_refs_complete_out_of_order(self, monkeypatch):
-        cluster = self._make_cluster()
-        port_finder = self.FakePortFinder()
-        results_by_ref = {
-            "ref-0": ("node-a", 25001),
-            "ref-1": ("node-a", 25002),
-            "ref-2": ("node-b", 25003),
-        }
-
-        def fake_wait(remaining, num_returns):
-            ready = remaining[-num_returns:]
-            remaining = remaining[:-num_returns]
-            return ready, remaining
-
-        self._patch_common_mocks(monkeypatch, port_finder, results_by_ref)
-        monkeypatch.setattr(virtual_cluster.ray, "wait", fake_wait)
-
-        results = cluster.get_available_addresses_and_ports_batch(
-            [(0, 0), (0, 1), (1, 0)], batch_size=1
-        )
-
-        assert results == [
-            ("node-a", 25001),
-            ("node-a", 25002),
-            ("node-b", 25003),
-        ]
-        assert port_finder.refs == [
-            ("ref-0", cluster.port_range_low, cluster.port_range_high),
-            ("ref-1", cluster.port_range_low, cluster.port_range_high),
-            ("ref-2", cluster.port_range_low, cluster.port_range_high),
-        ]
 
 
 class TestRayVirtualClusterPortRange:
